+++ UniMarket-Swift/UniMarket-Swift/Features/MainTab/MainTabView.swift
@@ -7,7 +7,7 @@
 
 import SwiftUI
 
-enum MainTab {
+enum MainTab: Hashable {
     case home
     case search
     case activity
@@ -18,27 +18,36 @@ struct MainTabView: View {
     @State private var selectedTab: MainTab = .home
     @State private var showUpload = false
 
+    private let barHeight: CGFloat = 64
+    private let sidePadding: CGFloat = 16
+    private let bottomSpacing: CGFloat = 0
+
     var body: some View {
-        TabHost(selectedTab: $selectedTab, showUpload: $showUpload)
-            .safeAreaInset(edge: .bottom, spacing: 0) {
+        GeometryReader { proxy in
+            ZStack {
+                tabContent
+                    .frame(width: proxy.size.width, height: proxy.size.height)
+
                 CustomTabBar(
                     selectedTab: $selectedTab,
                     onTapUpload: { showUpload = true }
                 )
-                .padding(.horizontal, 16)
-                .padding(.bottom, 8) // si la quieres MÁS abajo, pon 0
+                .frame(width: proxy.size.width - (sidePadding * 2), height: barHeight)
+                .position(
+                    x: proxy.size.width / 2,
+                    y: proxy.size.height - (barHeight / 2) - proxy.safeAreaInsets.bottom - bottomSpacing
+                )
             }
-            .sheet(isPresented: $showUpload) {
-                NavigationStack { UploadProductView() }
-            }
+            .frame(width: proxy.size.width, height: proxy.size.height)
+        }
+        .ignoresSafeArea(.keyboard, edges: .bottom)
+        .sheet(isPresented: $showUpload) {
+            NavigationStack { UploadProductView() }
+        }
     }
-}
 
-private struct TabHost: View {
-    @Binding var selectedTab: MainTab
-    @Binding var showUpload: Bool
-
-    var body: some View {
+    @ViewBuilder
+    private var tabContent: some View {
         switch selectedTab {
         case .home:
             NavigationStack {
@@ -54,61 +63,5 @@ private struct TabHost: View {
         case .profile:
             NavigationStack { ProfileView() }
         }
-    }
-}
-
-struct CustomTabBar: View {
-    @Binding var selectedTab: MainTab
-    let onTapUpload: () -> Void
-
-    var body: some View {
-        ZStack {
-            RoundedRectangle(cornerRadius: 18)
-                .fill(.background)
-                .shadow(radius: 10)
-                .frame(height: 64)
-
-            HStack {
-                tabButton(.home, icon: "house")
-                Spacer()
-                tabButton(.search, icon: "magnifyingglass")
-                Spacer()
-                Color.clear.frame(width: 56)
-                Spacer()
-                tabButton(.activity, icon: "heart")
-                Spacer()
-                tabButton(.profile, icon: "person")
-            }
-            .padding(.horizontal, 20)
-
-            Button {
-                onTapUpload()
-            } label: {
-                ZStack {
-                    Circle()
-                        .fill(.background)
-                        .shadow(radius: 10)
-                        .frame(width: 64, height: 64)
-                    Circle()
-                        .fill(Color.green.opacity(0.30))
-                        .frame(width: 56, height: 56)
-                    Image(systemName: "plus")
-                        .font(.system(size: 22, weight: .bold))
-                }
-            }
-            .offset(y: -18) // prueba -10 si lo quieres menos “alto”
-        }
-    }
-
-    private func tabButton(_ tab: MainTab, icon: String) -> some View {
-        Button {
-            selectedTab = tab
-        } label: {
-            Image(systemName: icon)
-                .font(.system(size: 20))
-                .foregroundStyle(selectedTab == tab ? .primary : .secondary)
-                .frame(width: 44, height: 44)
-        }
-        .buttonStyle(.plain)
     }
 }
